ai-service/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
google-genai==1.69.0
python-dotenv==1.2.2
ollama
